third_party/blink/web_tests/platform/generic/editing/inserting/5607069-3-expected.txt

This tests for a bug in InsertParagraph where it would split and clone a <form> element in the same way it would a <p> or a <div>. There should be exactly one form element and one br element between two input elements.
| <form>
|   style="border: 1px solid blue;"
|   <input>
|     type="text"
|     this.value=""
|     <shadow:root>
|       <div>
|   "x"
|   <br>
|   "<#selection-caret>x"
|   <input>
|     type="text"
|     this.value=""
|     <shadow:root>
|       <div>
